+++ smarts/core/sensors/__init__.py
@@ -24,7 +24,7 @@ from typing import Any, Dict, List, Optional, Set, Tuple
 
 import numpy as np
 
-from smarts.core.agent_interface import ActorsAliveDoneCriteria, AgentsAliveDoneCriteria
+from smarts.core.agent_interface import AgentsAliveDoneCriteria, InterestDoneCriteria
 from smarts.core.coordinates import Heading, Point
 from smarts.core.events import Events
 from smarts.core.observations import (
@@ -41,6 +41,7 @@ from smarts.core.observations import (
     Vias,
 )
 from smarts.core.plan import Plan, PlanFrame
+from smarts.core.renderer_base import RendererBase
 from smarts.core.road_map import RoadMap
 from smarts.core.sensor import (
     AccelerometerSensor,
@@ -70,7 +71,12 @@ ROAD_ID_CONSTANT = "off_road"
 LANE_INDEX_CONSTANT = -1
 
 
-def _make_vehicle_observation(road_map, neighborhood_vehicle):
+def _make_vehicle_observation(
+    road_map,
+    neighborhood_vehicle: VehicleState,
+    sim_frame: SimulationFrame,
+    interest_extension: Optional[re.Pattern],
+):
     nv_lane = road_map.nearest_lane(neighborhood_vehicle.pose.point, radius=3)
     if nv_lane:
         nv_road_id = nv_lane.road.road_id
@@ -91,6 +97,9 @@ def _make_vehicle_observation(road_map, neighborhood_vehicle):
         lane_id=nv_lane_id,
         lane_index=nv_lane_index,
         lane_position=None,
+        interest=sim_frame.actor_is_interest(
+            neighborhood_vehicle.actor_id, extension=interest_extension
+        ),
     )
 
 
@@ -102,14 +111,24 @@ class SensorState:
         self._plan_frame = plan_frame
         self._step = 0
         self._seen_interest_actors = False
+        self._seen_alive_actors = False
 
     def step(self):
         """Update internal state."""
         self._step += 1
 
     @property
+    def seen_alive_actors(self) -> bool:
+        """If an agents alive actor has been spotted before."""
+        return self._seen_alive_actors
+
+    @seen_alive_actors.setter
+    def seen_alive_actors(self, value: bool):
+        self._seen_alive_actors = value
+
+    @property
     def seen_interest_actors(self) -> bool:
-        """If a relevant actor has been spotted before."""
+        """If an interest actor has been spotted before."""
         return self._seen_interest_actors
 
     @seen_interest_actors.setter
@@ -143,7 +162,7 @@ class SensorResolver:
         sim_frame: SimulationFrame,
         sim_local_constants: SimulationLocalConstants,
         agent_ids: Set[str],
-        renderer,
+        renderer: RendererBase,
         bullet_client,
     ):
         """Generate observations
@@ -152,7 +171,7 @@ class SensorResolver:
             sim_frame (SimulationFrame): The simulation frame.
             sim_local_constants (SimulationLocalConstants): Constraints defined by the local simulator.
             agent_ids (Set[str]): The agents to run.
-            renderer (Renderer): The renderer to use.
+            renderer (RendererBase): The renderer to use.
             bullet_client (Any): The bullet client. This parameter is likely to be removed.
         """
         raise NotImplementedError()
@@ -259,10 +278,19 @@ class Sensors:
         )
         if neighborhood_vehicle_states_sensor:
             neighborhood_vehicle_states = []
+            interface = sim_frame.agent_interfaces.get(agent_id)
+            interest_pattern = (
+                interface.done_criteria.interest.actors_pattern
+                if interface is not None
+                and interface.done_criteria.interest is not None
+                else None
+            )
             for nv in neighborhood_vehicle_states_sensor(
                 vehicle_state, sim_frame.vehicle_states.values()
             ):
-                veh_obs = _make_vehicle_observation(sim_local_constants.road_map, nv)
+                veh_obs = _make_vehicle_observation(
+                    sim_local_constants.road_map, nv, sim_frame, interest_pattern
+                )
                 lane_position_sensor = vehicle_sensors.get("lane_position_sensor")
                 nv_lane_pos = None
                 if veh_obs.lane_id is not LANE_ID_CONSTANT and lane_position_sensor:
@@ -511,28 +539,17 @@ class Sensors:
         return False
 
     @classmethod
-    def _actors_alive_done_check(
+    def _interest_done_check(
         cls,
-        vehicle_ids,
-        sensor_state,
-        actors_alive: Optional[ActorsAliveDoneCriteria],
+        interest_actors,
+        sensor_state: SensorState,
+        interest_criteria: Optional[InterestDoneCriteria],
     ):
-        if actors_alive is None:
+        if interest_criteria is None or len(interest_actors) > 0:
+            sensor_state.seen_alive_actors = True
             return False
 
-        sensor_state: SensorState = sensor_state
-
-        pattern = re.compile(
-            "|".join(rf"(?:{aoi})" for aoi in actors_alive.actors_of_interest)
-        )
-        ## TODO optimization to get vehicles that were added and removed last step
-        ## TODO second optimization to check for already known vehicles
-        for vehicle_id in vehicle_ids:
-            # get vehicles by pattern
-            if pattern.match(vehicle_id):
-                sensor_state.seen_interest_actors = True
-                return False
-        if actors_alive.strict or sensor_state.seen_interest_actors:
+        if interest_criteria.strict or sensor_state.seen_interest_actors:
             # if agent requires the actor to exist immediately
             # OR if previously seen relevant actors but no actors match anymore
             return True
@@ -555,6 +572,7 @@ class Sensors:
         interface = sim_frame.agent_interfaces[agent_id]
         done_criteria = interface.done_criteria
         event_config = interface.event_configuration
+        interest = interface.done_criteria.interest
 
         # TODO:  the following calls nearest_lanes (expensive) 6 times
         reached_goal = cls._agent_reached_goal(
@@ -580,9 +598,13 @@ class Sensors:
         agents_alive_done = cls._agents_alive_done_check(
             sim_frame.ego_ids, sim_frame.potential_agent_ids, done_criteria.agents_alive
         )
-        actors_alive_done = cls._actors_alive_done_check(
-            sim_frame.vehicle_ids, sensor_state, done_criteria.actors_alive
-        )
+        interest_done = False
+        if interest:
+            cls._interest_done_check(
+                sim_frame.interest_actors(interest.actors_pattern),
+                sensor_state,
+                interest_criteria=interest,
+            )
 
         done = not sim_frame.resetting and (
             (is_off_road and done_criteria.off_road)
@@ -594,7 +616,7 @@ class Sensors:
             or (is_off_route and done_criteria.off_route)
             or (is_wrong_way and done_criteria.wrong_way)
             or agents_alive_done
-            or actors_alive_done
+            or interest_done
         )
 
         events = Events(
@@ -607,7 +629,7 @@ class Sensors:
             wrong_way=is_wrong_way,
             not_moving=is_not_moving,
             agents_alive_done=agents_alive_done,
-            actors_alive_done=actors_alive_done,
+            interest_done=interest_done,
         )
 
         return done, events