+++ smarts/core/smarts.py
@@ -20,6 +20,7 @@
 import importlib.resources as pkg_resources
 import logging
 import os
+import re
 import warnings
 from typing import Any, Dict, Iterable, List, Optional, Sequence, Set, Tuple, Union
 
@@ -32,6 +33,7 @@ from smarts import VERSION
 from smarts.core.actor_capture_manager import ActorCaptureManager
 from smarts.core.id_actor_capture_manager import IdActorCaptureManager
 from smarts.core.plan import Plan
+from smarts.core.renderer_base import RendererBase
 from smarts.core.simulation_local_constants import SimulationLocalConstants
 from smarts.core.utils.logging import timeit
 
@@ -76,6 +78,7 @@ logging.basicConfig(
     level=logging.ERROR,
 )
 
+_DEFAULT_PATTERN = re.compile("")
 MAX_PYBULLET_FREQ = 240
 
 
@@ -124,7 +127,7 @@ class SMARTS(ProviderManager):
         self._is_setup = False
         self._is_destroyed = False
         self._scenario: Optional[Scenario] = None
-        self._renderer = None
+        self._renderer: RendererBase = None
         self._envision: Optional[EnvisionClient] = envision
         self._visdom: Optional[VisdomClient] = visdom
         self._external_provider: ExternalProvider = None
@@ -217,6 +220,7 @@ class SMARTS(ProviderManager):
         Dict[str, Dict[str, float]],
     ]:
         """Progress the simulation by a fixed or specified time.
+
         Args:
             agent_actions:
                 Actions that the agents want to perform on their actors.
@@ -407,14 +411,15 @@ class SMARTS(ProviderManager):
         """Reset the simulation, reinitialize with the specified scenario. Then progress the
          simulation up to the first time an agent returns an observation, or `start_time` if there
          are no agents in the simulation.
+
         Args:
-            scenario(Scenario):
-                The scenario to reset the simulation with.
+            scenario(smarts.core.scenario.Scenario): The scenario to reset the simulation with.
             start_time(float):
                 The initial amount of simulation time to skip. This has implications on all time
                 dependent systems. NOTE: SMARTS simulates a step and then updates vehicle control.
                 If you want a vehicle to enter at exactly `0.3` with a step of `0.1` it means the
                 simulation should start at `start_time==0.2`.
+
         Returns:
             Agent observations. This observation is as follows:
                 - If no agents: the initial simulation observation at `start_time`
@@ -625,7 +630,7 @@ class SMARTS(ProviderManager):
         ), f"Vehicle has already been hijacked: {vehicle_id}"
         assert not vehicle_id in self.vehicle_index.agent_vehicle_ids(), (
             f"`{agent_id}` can't hijack vehicle that is already controlled by an agent"
-            f" `{self.vehicle_index.actor_id_from_vehicle_id(vehicle_id)}`: {vehicle_id}"
+            f" `{self.agent_manager.agent_for_vehicle(vehicle_id)}`: {vehicle_id}"
         )
 
         # Switch control to agent
@@ -957,13 +962,13 @@ class SMARTS(ProviderManager):
         )
 
     @property
-    def renderer(self):
+    def renderer(self) -> RendererBase:
         """The renderer singleton. On call, the sim will attempt to create it if it does not exist."""
         if not self._renderer:
             from .utils.custom_exceptions import RendererException
 
             try:
-                from .renderer import Renderer
+                from smarts.p3d.renderer import Renderer
 
                 self._renderer = Renderer(self._sim_id)
             except ImportError as e:
@@ -1053,10 +1058,10 @@ class SMARTS(ProviderManager):
         return VERSION
 
     def teardown_social_agents(self, agent_ids: Iterable[str]):
-        """
-        Teardown agents in the given sequence
-        Params:
-            agent_ids: Sequence of agent ids
+        """Teardown agents in the given sequence.
+
+        Args:
+            agent_ids: A sequence of agent ids to terminate and release.
         """
         agents_to_teardown = {
             id_
@@ -1067,10 +1072,10 @@ class SMARTS(ProviderManager):
         self.agent_manager.teardown_social_agents(filter_ids=agents_to_teardown)
 
     def teardown_social_agents_without_actors(self, agent_ids: Iterable[str]):
-        """
-        Teardown agents in the given list that have no actors registered as
+        """Teardown agents in the given list that have no actors registered as
         controlled-by or shadowed-by (for each given agent.)
-        Params:
+
+        Args:
             agent_ids: Sequence of agent ids
         """
         self._check_valid()
@@ -1381,7 +1386,6 @@ class SMARTS(ProviderManager):
     ) -> List[VehicleState]:
         """Find vehicles in the vicinity of the target vehicle."""
         self._check_valid()
-        from smarts.core.sensors import Sensors
 
         vehicle = self._vehicle_index.vehicle_by_id(vehicle_id)
         return neighborhood_vehicles_around_vehicle(
@@ -1680,4 +1684,7 @@ class SMARTS(ProviderManager):
             vehicle_sensors=self.sensor_manager.sensors_for_actor_ids(vehicle_ids),
             sensor_states=dict(self.sensor_manager.sensor_states_items()),
             _ground_bullet_id=self._ground_bullet_id,
+            interest_filter=self.scenario.metadata.get(
+                "actor_of_interest_re_filter", _DEFAULT_PATTERN
+            ),
         )