+++ smarts/diagnostic/10_agents_to_n_roads/1_roads/scenario.py
@@ -24,20 +24,22 @@ from pathlib import Path
 from smarts.sstudio import gen_scenario
 from smarts.sstudio.types import Mission, RandomRoute, Route, Scenario, SocialAgentActor
 
-actors = [
-    SocialAgentActor(
-        name=f"non-interactive-agent-50-v0",
-        agent_locator="zoo.policies:non-interactive-agent-v0",
-        policy_kwargs={"speed": 50},
-    )
-]
+
+def gen_actors(id_):
+    return [
+        SocialAgentActor(
+            name=f"non-interactive-agent-50-v0_{id_}",
+            agent_locator="zoo.policies:non-interactive-agent-v0",
+            policy_kwargs={"speed": 50},
+        )
+    ]
 
 
 def to_missions(agent_num):
     missions = {}
     for i in range(0, agent_num):
         missions[f"group-{i}"] = tuple(
-            (actors, [Mission(route=RandomRoute())]),
+            (gen_actors(i), [Mission(route=RandomRoute())]),
         )
     return missions