+++ smarts/env/gymnasium/platoon_env.py
@@ -25,9 +25,9 @@ from typing import Optional
 from envision.client import Client as Envision
 from envision.client import EnvisionDataFormatterArgs
 from smarts.core.agent_interface import (
-    ActorsAliveDoneCriteria,
     AgentInterface,
     DoneCriteria,
+    InterestDoneCriteria,
     NeighborhoodVehicles,
     Waypoints,
 )
@@ -58,12 +58,12 @@ def platoon_env(
     sumo_headless: bool = True,
     envision_record_data_replay_path: Optional[str] = None,
 ):
-    """All egos should track and follow the leader (i.e., lead vehicle) in a 
+    """All ego agents should track and follow the leader (i.e., lead vehicle) in a 
     single-file fashion. The lead vehicle is marked as a vehicle of interest 
     and may be found by filtering the 
     :attr:`~smarts.core.observations.VehicleObservation.interest` attribute of
     the neighborhood vehicles in the observation. The episode ends when the 
-    leader reaches its destination. Egos do not have prior knowledge of the 
+    leader reaches its destination. Ego agents do not have prior knowledge of the 
     leader's destination.
 
     Observation space for each agent:
@@ -173,8 +173,8 @@ def resolve_agent_interface(agent_interface: AgentInterface):
         on_shoulder=False,
         wrong_way=False,
         not_moving=False,
-        actors_alive=ActorsAliveDoneCriteria(
-            actors_of_interest=("Leader-007",),
+        interest=InterestDoneCriteria(
+            actors_filter=("Leader-007",),
             strict=True,
         ),
     )