+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Copy and pasting the git commit messages is __NOT__ enough.
 - The trap manager, `TrapManager`, is now a subclass of `ActorCaptureManager`.
 - Considering lane-change time ranges between 3s and 6s, assuming a speed of 13.89m/s, the via sensor lane acquisition range was increased from 40m to 80m, for better driving ability.
 - Modified naming of benchmark used in NeurIPS 2022 from driving-smarts-competition-env to driving-smarts-v2022.
+- Social agent actor vehicles are now exactly named the same as the `name` of the actor. 
 - Sstudio generated scenario vehicle traffic ids are now shortened.
 - ChaseViaPoints zoo agent uses unconstrained path change command, instead of being constrained to [-1, 0, +1] path change commands used previously. 
 - Made the metrics module configurable by supplying parameters through a `Params` class.
@@ -53,7 +54,9 @@ Copy and pasting the git commit messages is __NOT__ enough.
 - Benchmark listing may specify specialised metric formula for each benchmark.
 - Changed `benchmark_runner_v0.py` to only average records across scenarios that share the same environment. Records are not averaged across different environments, because the scoring formula may differ in different environments.
 - Renamed GapBetweenVehicles cost to VehicleGap cost in metric module.
-- Included `RelativeTargetPose` action space to set of allowed action spaces in `platoon-v0` env.
+- Camera metadata now uses radians instead of degrees.
+- The `Panda3d` implementation of `Renderer` has been extracted from the interface and moved to `smarts.p3d`.
+- Included `RelativeTargetPose` action space to the set of allowed action spaces in `platoon-v0` env.
 ### Deprecated
 ### Fixed
 - Fixed issues related to waypoints in junctions on Argoverse maps. Waypoints will now be generated for all paths leading through the lane(s) the vehicle is on.